requirements.txt
torch>=2.2
torchvision>=0.17
torchaudio>=2.2
datasets>=2.20
sentencepiece>=0.2
sacrebleu>=2.4
tokenizers>=0.19
loguru>=0.7
tensorboard>=2.16
tqdm>=4.66
pyyaml>=6.0